refactor(buildings): refactor building count into building state statistics
