+++ src/main/java/com/example/datn_qlnt_manager/service/BuildingService.java
@@ -20,7 +20,7 @@ public interface BuildingService {
 
     void deleteBuildingById(String buildingId);
 
-    BuildingCountResponse countBuildingByStatus ();
+    BuildingCountResponse statisticsBuildingByStatus();
 
     void toggleStatus(String id);
 }
